Added proof of work function

+++ src/main/kotlin/blockchain/domain/Blockchain.kt
@@ -5,7 +5,7 @@ typealias Blockchain = List<Block>
 data class Block(
     val index: Int,
     val timestamp: Long,
-    val proof: Int,
+    val proof: Double,
     val previousHash: String
 )
 
@@ -14,11 +14,11 @@ fun emptyBlockchain(timestamp: Long) = listOf(genesisBlock(timestamp))
 fun genesisBlock(timestamp: Long) = Block(
     index = 0,
     timestamp = timestamp,
-    proof = 1,
+    proof = 1.0,
     previousHash = "0"
 )
 
-fun Blockchain.createBlock(timestamp: Long, proof: Int, previousHash: String): Blockchain =
+fun Blockchain.createBlock(timestamp: Long, proof: Double, previousHash: String): Blockchain =
     Block(
         index = size,
         timestamp = timestamp,
